requirements.txt
# requirements.txt (make sure these are installed)
python-dotenv
langchain-google-genai
langchain-community
pinecone
pypdf
slack_bolt  # Slack framework
flask       # Web server integration for hosting
gunicorn    # Production-ready WSGI server (required for Render)
sentence-transformers